requirements.txt
pillow
opencv-python
numpy
cryptography